Refactor gradio frontend.py frontend.py refactor split into new files add ui_image_to_image.py add ui_text_to_image.py add ui_image_lab.py Refactor styles.css and custom.css Bugfix: Make the generate button not fill 1/2 the screen width
